server and client compile and conform to protocol, but explode

+++ common/include/serialization.h
@@ -2,7 +2,7 @@
 
 #include "shriek_types.h"
 
-serialized_message* alloc_serialized_message(const int connection_id,
+serialized_message* alloc_serialized_message(const ssize_t connection_id,
                                              char* const data,
                                              const size_t len);
 void free_serialized_message(serialized_message* s_message);
